promote scale to fp.
SVN revision: 35920

+++ legacy/evas/src/lib/canvas/evas_object_text.c
@@ -191,7 +191,7 @@ evas_object_text_font_set(Evas_Object *obj, const char *font, Evas_Font_Size siz
 	evas_font_free(obj->layer->evas, o->engine_data);
 	o->engine_data = NULL;
      }
-   o->engine_data = evas_font_load(obj->layer->evas, font, o->cur.source, size * obj->cur.scale);
+   o->engine_data = evas_font_load(obj->layer->evas, font, o->cur.source, (int)(((double)size) * obj->cur.scale));
    if (!same_font)
      {
 	if (o->cur.font) evas_stringshare_del(o->cur.font);
