+++ legacy/evas/src/lib/canvas/evas_object_textblock.c
@@ -1170,7 +1170,7 @@ _format_command(Evas_Object *obj, Evas_Object_Textblock_Format *fmt, const char 
 	  buf = strdup(fmt->font.name);
 	fmt->font.font = evas_font_load(obj->layer->evas, 
 					buf, fmt->font.source, 
-					fmt->font.size * obj->cur.scale);
+					(int)(((double)fmt->font.size) * obj->cur.scale));
 	if (buf) free(buf);
 	if (of) evas_font_free(obj->layer->evas, of);
      }
@@ -1291,7 +1291,7 @@ _format_dup(Evas_Object *obj, Evas_Object_Textblock_Format *fmt)
      buf = strdup(fmt2->font.name);
    fmt2->font.font = evas_font_load(obj->layer->evas, 
 				   buf, fmt2->font.source,
-				   fmt2->font.size * obj->cur.scale);
+				    (int)(((double)fmt2->font.size) * obj->cur.scale));
    if (buf) free(buf);
    return fmt2;
 }
